requirement.txt

django-crispy-forms 1.7.2